Adding options to import default duties
