+++ src/app/containers/ImportDefaultDutiesModal/ImportDefaultDutiesModal.css
@@ -1,0 +1,7 @@
+.import-default-toggle {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding-bottom: 15px;
+    font-size: 15px; 
+}
